+++ Tests/TuistGeneratorTests/Graph/TestData/CocoaPods+TestData.swift
@@ -1,0 +1,10 @@
+import Basic
+import Foundation
+
+@testable import TuistGenerator
+
+extension CocoaPodsNode {
+    static func test(path: AbsolutePath = "/") -> CocoaPodsNode {
+        return CocoaPodsNode(path: path)
+    }
+}
