+++ Tests/TuistGeneratorTests/Graph/TestData/FrameworkNode+TestData.swift
@@ -1,0 +1,10 @@
+import Basic
+import Foundation
+
+@testable import TuistGenerator
+
+extension FrameworkNode {
+    static func test(path: AbsolutePath = "/Test.framework") -> FrameworkNode {
+        return FrameworkNode(path: path)
+    }
+}
