+++ Tests/TuistGeneratorTests/Graph/TestData/TargetNode+TestData.swift
@@ -12,16 +12,3 @@ extension TargetNode {
                           dependencies: dependencies)
     }
 }
-
-extension FrameworkNode {
-    static func test(path: AbsolutePath = "/Test.framework") -> FrameworkNode {
-        return FrameworkNode(path: path)
-    }
-}
-
-extension LibraryNode {
-    static func test(path: AbsolutePath = "/libTest.a",
-                     publicHeaders: AbsolutePath = "/TestHeaders/") -> LibraryNode {
-        return LibraryNode(path: path, publicHeaders: publicHeaders)
-    }
-}
